Split physics part from objects file
